symbia-app/Server/requirements.txt
langchain
langchain-openai
langchain-community
langchain-text-splitters
faiss-cpu
beautifulsoup4
python-dotenv
requests
fastapi
uvicorn
